+++ src/main/java/com/example/aitalk/config/Response.java
@@ -1,4 +1,4 @@
-package com.example.aitalk.security;
+package com.example.aitalk.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
