+++ src/main/java/com/example/aitalk/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.aitalk.swagger;
+package com.example.aitalk.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
