+++ src/main/java/com/example/aitalk/member/MemberController.java
@@ -1,6 +1,6 @@
 package com.example.aitalk.member;
 
-import com.example.aitalk.security.Response;
+import com.example.aitalk.config.Response;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
